requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
groq
python-multipart